requirements.txt
# Core dependencies
flask==2.3.3
ib_insync==0.9.86
python-telegram-bot==20.7
requests==2.31.0
schedule==1.2.0

# Data handling
pandas==2.1.4
numpy==1.24.3
pytz==2023.3

# Configuration and logging
pyyaml==6.0.1
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Additional utilities
colorama==0.4.6
rich==13.7.0